requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
oracledb==3.4.1
python-dotenv==1.0.1
python-multipart==0.0.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
jinja2==3.1.5
httpx==0.28.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-httpx==0.50b0
opentelemetry-instrumentation-logging==0.50b0
opentelemetry-exporter-prometheus==0.50b0
prometheus-client==0.21.1
oci==2.143.0
pydantic==2.10.4
pydantic-settings==2.7.1
itsdangerous==2.2.0
lxml==5.3.0
pyyaml==6.0.2
